added chatroom list

+++ notgitter/Models/Message.cs
@@ -18,7 +18,7 @@ namespace notgitter.Models
         public string Content { get; set; }
         public Nullable<System.DateTime> timestamp { get; set; }
         public string UserName { get; set; }
-        public long RepoId { get; set; }
+        public int RepoId { get; set; }
         public int Uid { get; set; }
     
         public virtual Repo Repo { get; set; }
